fix(playground): remove md2pdf template handoff actions
